fix: align model stream connection ownership
